refactor: package structure 를 세분화
