--- src/main/java/lotto/domain/LottoGenerator.java
@@ -1,5 +1,0 @@
-package lotto.domain;
-
-public interface LottoGenerator {
-	Lotto createLotto();
-}
