+++ src/main/java/lotto/domain/production/LottoGenerator.java
@@ -1,0 +1,7 @@
+package lotto.domain.production;
+
+import lotto.domain.lotto.Lotto;
+
+public interface LottoGenerator {
+	Lotto createLotto();
+}
