+++ src/main/java/lotto/domain/production/ManualLottoGenerator.java
@@ -1,4 +1,7 @@
-package lotto.domain;
+package lotto.domain.production;
+
+import lotto.domain.lotto.Lotto;
+import lotto.domain.lotto.LottoNumber;
 
 import java.util.ArrayList;
 import java.util.List;
