+++ src/main/java/lotto/view/InputView.java
@@ -1,8 +1,7 @@
 package lotto.view;
 
-import lotto.domain.LottoMachine;
+import lotto.domain.production.LottoMachine;
 
-import java.io.InputStream;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
