+++ src/test/java/lotto/domain/lotto/LottoTest.java
@@ -1,13 +1,14 @@
-package lotto.domain;
+package lotto.domain.lotto;
 
+import lotto.domain.result.WinningLotto;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import java.util.Arrays;
 
-import static lotto.domain.Lotto.DUPLICATE_ERROR;
-import static lotto.domain.Lotto.LOTTO_SIZE_ERROR;
+import static lotto.domain.lotto.Lotto.DUPLICATE_ERROR;
+import static lotto.domain.lotto.Lotto.LOTTO_SIZE_ERROR;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
