+++ src/test/java/lotto/domain/production/AutomaticLottoGeneratorTest.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.production;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
