+++ src/test/java/lotto/domain/production/LottoMachineTest.java
@@ -1,5 +1,9 @@
-package lotto.domain;
+package lotto.domain.production;
 
+import lotto.domain.Money;
+import lotto.domain.lotto.Lotto;
+import lotto.domain.lotto.LottoNumber;
+import lotto.domain.lotto.Lottos;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
