+++ src/test/java/lotto/domain/production/ManualLottoGeneratorTest.java
@@ -1,5 +1,7 @@
-package lotto.domain;
+package lotto.domain.production;
 
+import lotto.domain.lotto.Lotto;
+import lotto.domain.lotto.LottoNumber;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
