+++ src/test/java/lotto/domain/result/LottoStatisticsTest.java
@@ -1,5 +1,9 @@
-package lotto.domain;
+package lotto.domain.result;
 
+import lotto.domain.Money;
+import lotto.domain.lotto.Lotto;
+import lotto.domain.lotto.LottoNumber;
+import lotto.domain.lotto.Lottos;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
