+++ src/test/java/lotto/domain/result/RankTest.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.result;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
