+++ src/test/java/lotto/view/OutputViewTest.java
@@ -1,6 +1,13 @@
 package lotto.view;
 
-import lotto.domain.*;
+import lotto.domain.Money;
+import lotto.domain.lotto.Lotto;
+import lotto.domain.lotto.LottoNumber;
+import lotto.domain.lotto.Lottos;
+import lotto.domain.production.LottoMachine;
+import lotto.domain.production.LottoQuantity;
+import lotto.domain.result.LottoStatistics;
+import lotto.domain.result.Rank;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
